+++ database/migrations/2021_12_05_000003_create_subscriptions_table.php
@@ -1,8 +1,9 @@
 <?php
 
-use Illuminate\Database\Migrations\Migration;
-use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+use FintechSystems\Payfast\Enums\PaymentMethod;
 
 class CreateSubscriptionsTable extends Migration
 {
@@ -17,14 +18,26 @@ class CreateSubscriptionsTable extends Migration
             $table->id();
             $table->unsignedBigInteger('billable_id');
             $table->string('billable_type');
-            $table->string('name');
             $table->integer('plan_id');
+            $table->string('name');
+                        
+            $table->unsignedBigInteger('amount')->nullable();
+            $table->unsignedInteger('cycles')->nullable();
+            $table->unsignedInteger('cycles_complete')->nullable();
+            $table->unsignedInteger('frequency')->nullable();
+            $table->unsignedInteger('payfast_status')->nullable();
+            $table->string('payfast_status_reason')->nullable();
+            $table->string('payfast_status_text')->nullable();
             $table->string('token')->unique();
+
+            $table->string('payment_method');
+
             $table->string('status');
             $table->string('merchant_payment_id');
             $table->string('payment_status')->nullable();
             $table->string('subscription_status')->nullable();
             $table->timestamp('next_bill_at')->nullable();
+            $table->timestamp('next_run_at')->nullable();
             $table->timestamp('ends_at')->nullable();
             $table->timestamp('cancelled_at')->nullable();
             $table->timestamp('trial_ends_at')->nullable();